// admin-fe/src/app/model/interface/phieu-update.interface.ts
export interface PhieuGiamGiaUpdate {
    id?: number;
    maPhieuGiamGia?: string;
    tenPhieuGiamGia?: string;
    kieu?: number;
    loai?: number;
    giaTri?: number;
    giaTriMax?: number;
    dieuKienGiam?: number;
    soLuong?: number;
    thoiGianBatDau?: string;
    thoiGianKetThuc?: string;
    trangThai?: string;
}